requirements.txt
flask==2.3.3
flask-session==0.5.0
python-telegram-bot==20.7
gunicorn==22.0.0
psycopg2-binary==2.9.9